fix: rename participationHelper to participationUtils :)
